Extracted multiplex creation in separate helper
